Started work on refactoring Task into a proper part of the library.
